[Ripple][Clang] Fix lower bound type re-cast
When starting with an implicit cast, applying another implicit cast in
clang can override the first one. However, the initial expression needs
the original semantics (implicit cast). The solution is to add an extra
parenthesis expression so that the implicit cast method will not modify
the pre-existing implicit casts.

This code comes from OpenMP but the but does not apply to it because the
induction variable is private. For Ripple, we reuse the induction
variable and its init value.

+++ clang/test/Ripple/ripple-parallel/thread-thread-nesting-32bits.c
@@ -236,8 +236,9 @@
 // CHECK:       [[RIPPLE_PAR_FOR_END67]]:
 // CHECK-NEXT:    ret void
 //
-void check(int Chunk, int64_t N, int32_t start, int64_t end, float x[restrict N],
-           float y[restrict N], float xpy[restrict N]) {
+void check(int Chunk, long long N, long start, long long end,
+           float x[restrict N], float y[restrict N],
+           float xpy[restrict N]) {
   ripple_block_t BST = ripple_set_block_shape(32, 32);
   ripple_block_t BSV = ripple_set_block_shape(33, 4, 4);
   #ifdef USING_PRAGMA
